+++ src/Core/Modules/CONSOLE/ConsoleCommandReply.php
@@ -265,10 +265,35 @@ class ConsoleCommandReply implements CommandReply, MessageEmitter {
 			},
 			$message
 		);
+		$useHyperlinks = true;
 		$message = Safe::pregReplace("/<a\s+href=['\"]?user:\/\/[^'\">]+['\"]?\s*>(.*?)<\/a>/s", '<link>$1</link>', $message);
 		$message = Safe::pregReplace("/<a\s+href=['\"]?skillid:\/\/\d+['\"]?\s*>(.*?)<\/a>/s", '[skill:<link>$1</link>]', $message);
 		$message = Safe::pregReplace("/<a\s+href=['\"]chatcmd:\/\/\/(.*?)['\"]\s*>(.*?)<\/a>/s", '<link>$2</link>', $message);
-		$message = Safe::pregReplace("/<a\s+href=['\"]?itemref:\/\/\d+\/\d+\/\d+['\"]?\s*>(.*?)<\/a>/s", '[item:<link>$1</link>]', $message);
+		$useHyperlinks = $this->consoleController->consoleItemDisplay !== $this->consoleController::PLACEHOLDERS;
+		if ($useHyperlinks) {
+			$message = Safe::pregReplaceCallback(
+				"/<a\s+href=['\"]?itemref:\/\/(\d+)\/\d+\/(\d+)['\"]?\s*>(.*?)<\/a>/s",
+				function (array $matches): string {
+					$schema = $this->consoleController->consoleItemDisplay;
+					$url = Text::renderPlaceholders($schema, ['id' => $matches[1], 'ql' => $matches[2]]);
+					return $this->createLink($url, $matches[3]);
+				},
+				$message
+			);
+			$message = Safe::pregReplaceCallback(
+				"/<a\s+href=['\"]?itemid:\/\/53019\/(\d+)['\"]?\s*>(.*?)<\/a>/s",
+				function (array $matches): string {
+					$schema = $this->consoleController->consoleItemDisplay;
+					$url = Text::renderPlaceholders($schema, ['id' => $matches[1]]);
+					return $this->createLink($url, $matches[2]);
+				},
+				$message
+			);
+			$message = Safe::pregReplace("/<a\s+href=['\"]?itemid:\/\/53019\/\d+['\"]?\s*>(.*?)<\/a>/s", '[nano:<link>$1</link>]', $message);
+		} else {
+			$message = Safe::pregReplace("/<a\s+href=['\"]?itemref:\/\/\d+\/\d+\/\d+['\"]?\s*>(.*?)<\/a>/s", '[item:<link>$1</link>]', $message);
+			$message = Safe::pregReplace("/<a\s+href=['\"]?itemid:\/\/53019\/\d+['\"]?\s*>(.*?)<\/a>/s", '[nano:<link>$1</link>]', $message);
+		}
 		$message = Safe::pregReplace("/<a\s+href=['\"]?itemid:\/\/53019\/\d+['\"]?\s*>(.*?)<\/a>/s", '[nano:<link>$1</link>]', $message);
 		$message = Safe::pregReplace("/<p\s*>/is", "\n", $message);
 		$message = Safe::pregReplace("/<\/p\s*>/is", '', $message);
@@ -409,5 +434,9 @@ class ConsoleCommandReply implements CommandReply, MessageEmitter {
 	protected function bgHexToAnsi(string $hexColor): string {
 		$codes = array_map('hexdec', str_split($hexColor, 2));
 		return "\e[48;2;" . implode(';', $codes) . 'm';
+	}
+
+	private function createLink(string $url, string $text): string {
+		return "<link>\e]8;;{$url}\e\\{$text}\e]8;;\e\\</link>";
 	}
 }